+++ src/posts/posts.service.ts
@@ -1,11 +1,19 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { PostsRepository } from './posts.repository';
+import { PublicationsRepository } from '../publications/publications.repository';
 
 @Injectable()
 export class PostsService {
-  constructor(private readonly postsRepository: PostsRepository) {}
+  constructor(
+    private readonly postsRepository: PostsRepository,
+    private readonly publicationsRepository: PublicationsRepository,
+  ) {}
 
   async create(createPostDto: CreatePostDto) {
     return await this.postsRepository.create(createPostDto);
@@ -38,6 +46,13 @@ export class PostsService {
     if (!postExists) {
       throw new NotFoundException('Theres no post with this id');
     }
-    return await this.postsRepository.remove(id);
+
+    const publicationExists = this.publicationsRepository.findOne(id);
+    if (publicationExists) {
+      throw new ForbiddenException('Theres a publication with this post!');
+    }
+
+    await this.postsRepository.remove(id);
+    return 'Post deleted!';
   }
 }
